[SCRUM-57] Timer Top-Down 형태로 수정 (#151)
