+++ presentation/src/main/java/com/pomonyang/mohanyang/presentation/screen/pomodoro/PomodoroTimerViewModel.kt
@@ -7,53 +7,17 @@ import com.pomonyang.mohanyang.data.repository.user.UserRepository
 import com.pomonyang.mohanyang.domain.usecase.GetSelectedPomodoroSettingUseCase
 import com.pomonyang.mohanyang.domain.usecase.InsertPomodoroInitialDataUseCase
 import com.pomonyang.mohanyang.presentation.base.BaseViewModel
-import com.pomonyang.mohanyang.presentation.base.ViewEvent
-import com.pomonyang.mohanyang.presentation.base.ViewSideEffect
-import com.pomonyang.mohanyang.presentation.base.ViewState
-import com.pomonyang.mohanyang.presentation.model.cat.CatType
 import com.pomonyang.mohanyang.presentation.model.cat.toModel
-import com.pomonyang.mohanyang.presentation.model.setting.PomodoroCategoryType
 import com.pomonyang.mohanyang.presentation.model.setting.toModel
 import com.pomonyang.mohanyang.presentation.screen.PomodoroConstants
-import com.pomonyang.mohanyang.presentation.util.formatTime
 import dagger.hilt.android.lifecycle.HiltViewModel
-import java.util.UUID
+import java.util.*
 import javax.inject.Inject
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.launch
-
-data class PomodoroTimerState(
-    val focusTime: Int = 0,
-    val focusExceededTime: Int = 0,
-    val maxFocusTime: Int = 0,
-    val restTime: Int = 0,
-    val restExceededTime: Int = 0,
-    val maxRestTime: Int = 0,
-    val title: String = "",
-    val categoryType: PomodoroCategoryType = PomodoroCategoryType.DEFAULT,
-    val cat: CatType = CatType.CHEESE,
-    val categoryNo: Int = -1,
-    val forceGoRest: Boolean = false
-) : ViewState {
-    fun displayFocusTime(): String = focusTime.formatTime()
-    fun displayRestTime(): String = restTime.formatTime()
-    fun displayFocusExceedTime(): String = focusExceededTime.formatTime()
-    fun displayRestExceedTime(): String = restExceededTime.formatTime()
-}
-
-sealed interface PomodoroTimerEvent : ViewEvent {
-    // 포커스, 휴식 이벤트 분리를 했는데 나중에 필요하면 상속받고 추가
-    sealed interface PomodoroFocusEvent : PomodoroTimerEvent
-    sealed interface PomodoroRestEvent : PomodoroTimerEvent
-}
-
-sealed interface PomodoroTimerEffect : ViewSideEffect {
-    sealed interface PomodoroFocusEffect : PomodoroTimerEffect
-    sealed interface PomodoroRestEffect : PomodoroTimerEffect
-}
 
 @HiltViewModel
 class PomodoroTimerViewModel @Inject constructor(
@@ -115,16 +79,16 @@ class PomodoroTimerViewModel @Inject constructor(
     }
 
     private fun updateTimerState(timerData: PomodoroTimerEntity) {
-        val currentFocusTime = timerData.focusedTime.coerceAtMost(state.value.maxFocusTime)
-        val currentRestTime = timerData.restedTime.coerceAtMost(state.value.maxRestTime)
+        val remainingFocusTime = (state.value.maxFocusTime - timerData.focusedTime).coerceAtLeast(0)
+        val remainingRestTime = (state.value.maxRestTime - timerData.restedTime).coerceAtLeast(0)
         val focusExceededTime = (timerData.focusedTime - state.value.maxFocusTime).coerceAtLeast(0)
         val restExceededTime = (timerData.restedTime - state.value.maxRestTime).coerceAtLeast(0)
 
         updateState {
             copy(
-                focusTime = currentFocusTime,
+                remainingFocusTime = remainingFocusTime,
                 focusExceededTime = focusExceededTime,
-                restTime = currentRestTime,
+                remainingRestTime = remainingRestTime,
                 restExceededTime = restExceededTime
             )
         }
